fix: inline setup token logic matching Twilio/Blooio pattern
Remove the separate onboarding-setup-token.ts module per review
feedback — the logic duplicated subscription-routes.ts and lacked
a runMode guard. Inline the token handling directly in the
subscription-provider block of the onboarding handler (already
guarded by runMode === "local"), matching how Twilio/Blooio tokens
are handled in the same handler. Add cross-reference comment to
subscription-routes.ts for the equivalent endpoint.
