+++ src/api/server.ts
@@ -97,7 +97,6 @@ import {
 } from "./http-helpers";
 import { handleKnowledgeRoutes } from "./knowledge-routes";
 import { handleModelsRoutes } from "./models-routes";
-import { applySubscriptionSetupToken } from "./onboarding-setup-token";
 import { handlePermissionRoutes } from "./permissions-routes";
 import {
   type PluginParamInfo,
@@ -4855,11 +4854,19 @@ async function handleRequest(
         `[milady-api] Subscription provider selected: ${body.provider} — complete OAuth via /api/subscription/ endpoints`,
       );
 
-      const setupResult = applySubscriptionSetupToken(
-        body,
-        config as { env?: Record<string, string> },
-      );
-      if (setupResult.saved) {
+      // Handle Anthropic setup token (sk-ant-oat01-...) provided during
+      // onboarding. The API-key gate above skips subscription providers
+      // because their envKey is null. Mirrors POST /api/subscription/
+      // anthropic/setup-token in subscription-routes.ts.
+      if (
+        body.provider === "anthropic-subscription" &&
+        typeof body.providerApiKey === "string" &&
+        body.providerApiKey.trim().startsWith("sk-ant-")
+      ) {
+        const token = body.providerApiKey.trim();
+        if (!config.env) config.env = {};
+        (config.env as Record<string, string>).ANTHROPIC_API_KEY = token;
+        process.env.ANTHROPIC_API_KEY = token;
         logger.info(
           "[milady-api] Anthropic setup token saved during onboarding",
         );
